feat(network): remove external address manager
